pyproject.toml
[build-system]
requires = ["setuptools >= 65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "app"
description = "Test task Fastapi"
readme = "README.md"
authors = [
    {name = "mr.marshanskiy"},
]
classifiers = [
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"

dependencies = [
    "click==8.1.7",
    "fastapi==0.104.1",
    "gunicorn==21.2.0",
    "loguru==0.7.2",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.4.2",
    "pydantic-settings==2.0.3",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy==2.0.23",
    "uvicorn[standard]==0.24.0.post1",
]

dynamic = ["version"]

[project.optional-dependencies]
check = [
    "black",
    "isort",
    "mypy",
]
test = [
    "httpx>=0.24.0",
    "pytest",
    "pytest-cov",
]

[project.scripts]
myapi = "app.cli:main"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.dynamic]
version = {attr = "app.version.__version__"}

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
lines_after_imports = 2
force_grid_wrap = 2
force_sort_within_sections = "True"
order_by_type = "False"

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = "True"
ignore_missing_imports = "True"
show_column_numbers = "True"
disallow_untyped_calls = "True"
strict_optional = "True"
warn_no_return = "True"
warn_redundant_casts = "True"
warn_unused_ignores = "True"
check_untyped_defs = "True"
install_types = "True"
non_interactive = "True"

[tool.pytest.ini_options]
addopts = "--color=yes --cov=app --cov-report=term-missing -vv -s"
testpaths = [
    "tests",
]